Refact class Services

+++ src/main/java/com/wipro/techbank/services/SpecialAccountService.java
@@ -1,11 +1,12 @@
 package com.wipro.techbank.services;
 
+import com.wipro.techbank.domain.CheckingAccount;
 import com.wipro.techbank.domain.Client;
+import com.wipro.techbank.domain.CreditCard;
 import com.wipro.techbank.domain.SpecialAccount;
-import com.wipro.techbank.dtos.SpecialAccountDto;
-import com.wipro.techbank.dtos.SpecialAccountRequestDto;
-import com.wipro.techbank.dtos.SpecialAccountResponseDto;
+import com.wipro.techbank.dtos.*;
 import com.wipro.techbank.repositories.ClientRepository;
+import com.wipro.techbank.repositories.CreditCardRepository;
 import com.wipro.techbank.repositories.SpecialAccountRepository;
 import com.wipro.techbank.services.exceptions.DataBasesException;
 import com.wipro.techbank.services.exceptions.ResourceNotFoundException;
@@ -18,6 +19,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -33,61 +35,81 @@ public class SpecialAccountService {
     @Autowired
     private ModelMapper modelMapper;
 
-    public List<SpecialAccountResponseDto> findAll(){
+    @Autowired
+    private CreditCardRepository creditCardRepository;
+
+    public List<SpecialAccountResponseDto> findAll() {
         return specialAccountRepository.findAll()
                 .stream()
                 .map(this::toSpecialAccountDto)
                 .collect(Collectors.toList());
     }
 
-    public Page<SpecialAccountDto> findAllPaged(Pageable pageable) {
-        Page<SpecialAccount> specialAccounts = specialAccountRepository.findAll(pageable);
-        return specialAccounts.map(SpecialAccountDto::new);
-    }
-
-    public SpecialAccountResponseDto findById(Long id){
+    public SpecialAccountResponseDto findById(Long id) {
         Optional<SpecialAccount> optionalSpecialAccount = specialAccountRepository.findById(id);
-        SpecialAccount specialAccountDb = optionalSpecialAccount.orElseThrow(()->
+        SpecialAccount specialAccountDb = optionalSpecialAccount.orElseThrow(() ->
                 new ResourceNotFoundException("Entidade não encontrada"));
         return toSpecialAccountDto(specialAccountDb);
     }
 
-    public SpecialAccountResponseDto create(SpecialAccountRequestDto specialAccountRequestDto){
+    public SpecialAccountResponseDto create(SpecialAccountRequestDto specialAccountRequestDto) {
         Long idClient = specialAccountRequestDto.getClient().getId();
         Optional<Client> client = clientRepository.findById(idClient);
-        if(!client.isPresent()){
+        if (!client.isPresent()) {
             throw new ResourceNotFoundException("Entidade não encontrada");
         }
         SpecialAccount specialAccount = toSpecialAccount(specialAccountRequestDto);
-        specialAccount = specialAccountRepository.save(specialAccount);
+        specialAccountRepository.save(specialAccount);
         return toSpecialAccountDto(specialAccount);
+
     }
 
-    public SpecialAccountResponseDto updateSpecialAccount(Long id, SpecialAccountRequestDto specialAccountRequestDto){
+    public SpecialAccountResponseDto updateSpecialAccount(Long id, SpecialAccountRequestDto specialAccountRequestDto) {
         Optional<SpecialAccount> optionalSpecialAccountDb = specialAccountRepository.findById(id);
         SpecialAccount specialAccountDb = specialAccountRepository.findById(id).orElseThrow(() ->
                 new ResourceNotFoundException("Entidade não econtrada."));
 
-        BeanUtils.copyProperties(specialAccountRequestDto,specialAccountDb);
+        BeanUtils.copyProperties(specialAccountRequestDto, specialAccountDb);
         specialAccountRepository.save(specialAccountDb);
         return toSpecialAccountDto(specialAccountDb);
     }
 
-    public void delete(Long id){
-        try {
+    public void remove(Long id) {
+        Optional<SpecialAccount> optionalSpecialAccount = specialAccountRepository.findById(id);
+        if (optionalSpecialAccount.isPresent()) {
             specialAccountRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new ResourceNotFoundException("Id " + id + " não encontrado.");
-        } catch (DataIntegrityViolationException e) {
-            throw new DataBasesException("Violação de integridade.");
+        } else {
+            throw new ResourceNotFoundException("Entidade não encontrada");
         }
     }
 
-    private SpecialAccountResponseDto toSpecialAccountDto(SpecialAccount specialAccount){
+    private SpecialAccountResponseDto toSpecialAccountDto(SpecialAccount specialAccount) {
         return modelMapper.map(specialAccount, SpecialAccountResponseDto.class);
     }
 
-    private SpecialAccount toSpecialAccount(SpecialAccountRequestDto specialAccountRequestDto){
+    private SpecialAccount toSpecialAccount(SpecialAccountRequestDto specialAccountRequestDto) {
+        if (specialAccountRequestDto.getClient().equals(null)) {
+            throw new ResourceNotFoundException("Cliente, cadastrado, precisa ser informado");
+        }
+
+        try {
+            Client client = clientRepository.getById(specialAccountRequestDto.getClient().getId());
+            specialAccountRequestDto.setClient(client);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException(String.format("Cliente com ID %d não foi encontrado.", specialAccountRequestDto.getClient().getId()));
+        }
+
+        if (specialAccountRequestDto.getCreditCard().equals(null)) {
+            throw new ResourceNotFoundException("Cartão de Credito, cadastrado, precisa ser informado.");
+        }
+
+        try {
+            CreditCard creditCard = creditCardRepository.getById(specialAccountRequestDto.getCreditCard().getId());
+            specialAccountRequestDto.setCreditCard(creditCard);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException(String.format("Cartão de Credito com ID %d não foi encontrado.", specialAccountRequestDto.getCreditCard().getId()));
+        }
+
         return modelMapper.map(specialAccountRequestDto, SpecialAccount.class);
     }
 }
